requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
docopt==0.6.2
Flask==1.1.2
future==0.18.2
geoserver-rest==1.2.4
geoserver-restconfig==2.0.4.8
gisdata==0.5.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Js2Py==0.70
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
packaging==20.7
pandas==1.1.4
Pillow==8.0.1
pipwin==0.5.0
psycopg2==2.8.6
pycurl==7.43.0.6
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
soupsieve==2.0.1
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
